Clean up
- Clear up documentation
- Add Ansible vars to CI
- Add views.py
