+++ deployment/ansible/group_vars/all
@@ -43,4 +43,4 @@ docker_version: "18.*"
 model_data_path: "/opt/icp-crop-data"
 numpy_version: "1.11.1"
 
-ubuntu_ssl_certs: "/etc/ssl/certs/ca-certificates.crt"
+ubuntu_ssl_cert_path: "/etc/ssl/certs/ca-certificates.crt"
