+++ deployment/ansible/group_vars/development
@@ -23,7 +23,7 @@ celery_processes_per_worker: 1
 
 docker_options: "-H tcp://0.0.0.0:2375 -H unix:///var/run/docker.sock --storage-driver=aufs"
 
-aws_profile: "icp_bees"
+aws_profile: "icp-bees"
 aws_beekeepers_data_bucket: "beekeepers-staging-data-us-east-1"
 
 stack_type: "Development"
